backend/requirements.txt
fastapi>=0.100.0
slowapi>=0.1.9
pydantic-settings>=2.0.0
# 文件上传（UploadFile / multipart 表单）依赖
python-multipart>=0.0.6
uvicorn[standard]>=0.22.0
PyJWT>=2.0.0
httpx>=0.27.0
SQLAlchemy>=2.0.0
aiomysql>=0.2.0
passlib[bcrypt]>=1.7.4
# passlib 与 bcrypt 4.1+ 不兼容，需限制版本
bcrypt>=4.0.1,<4.1
redis>=5.0.0
pandas>=2.0.0
openpyxl>=3.1.0
